backend/requirements.txt
APScheduler==3.10.4
Flask==3.0.3
Flask_Cors==5.0.0
numpy==2.1.3
opencv_python==4.10.0.84
Pillow==11.0.0
Requests==2.32.3
gunicorn==20.1.0
python-dotenv